+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
-### New in 0.11 (not released yet)
+### New in 0.12 (not released yet)
+ 
+ * _Nothing yet_
 
- * Breaking: `EventFlowOptions AddDefaults(...)` now also adds event
+### New in 0.11.751 (released 2015-08-24)
+
+ * Breaking: `EventFlowOptions.AddDefaults(...)` now also adds event
    definitions
  * New: [RabbitMQ](http://www.rabbitmq.com/) is now supported through the new
    NuGet package called `EventFlow.RabbitMQ` which enables domain events to be
@@ -9,6 +13,17 @@
    and register a service that implements `ISubscribeSynchronousToAll`. Services
    that implement this will automatically be added using the
    `AddSubscribers(...)` or `AddDefaults(...)` extension to `EventFlowOptions`
+ * New: Use `EventFlowOptions.UseAutofacAggregateRootFactory(...)` to use an
+   Autofac aggregate root factory, enabling you to use services in your
+   aggregate root constructor
+ * New: Use `EventFlowOptions.UseResolverAggregateRootFactory()` to use the
+   resolver to create aggregate roots. Same as
+   `UseAutofacAggregateRootFactory(...)` but for when using the internal IoC
+   container
+ * New: Use `EventFlowOptions.AddAggregateRoots(...)` to register aggregate root
+   types
+ * New: Use `IServiceRegistration.RegisterType(...)` to register services by
+   type
 
 ### New in 0.10.642 (released 2015-08-17)